style: cleaned code by creating file
