+++ app/src/main/java/com/github/se/assocify/ui/screens/treasury/accounting/AccountingDetailedScreen.kt
@@ -63,6 +63,7 @@ import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.composables.DatePickerWithDialog
 import com.github.se.assocify.ui.composables.DropdownFilterChip
 import com.github.se.assocify.ui.screens.treasury.accounting.balance.BalanceDetailedViewModel
+import com.github.se.assocify.ui.screens.treasury.accounting.balance.DisplayEditBalance
 import com.github.se.assocify.ui.screens.treasury.accounting.budget.BudgetDetailedViewModel
 import java.time.LocalDate
 
@@ -260,272 +261,6 @@ fun DisplayBalanceItem(
       Modifier
         .clickable { balanceDetailedViewModel.startEditing(balanceItem) }
         .testTag(testTag))
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DisplayEditBalance(balanceDetailedViewModel: BalanceDetailedViewModel) {
-  val balanceModel by balanceDetailedViewModel.uiState.collectAsState()
-  val balance = balanceModel.editedBalanceItem!!
-  var nameString by remember { mutableStateOf("") }
-  var subCategoryUid by remember { mutableStateOf("") }
-  var subCategoryName by remember { mutableStateOf("") }
-  var receiptUid by remember { mutableStateOf("") }
-  var receiptName by remember { mutableStateOf("") }
-  var amount by remember { mutableIntStateOf(0) }
-  var tvaString by remember { mutableStateOf("") }
-  var tvaTypeString by remember { mutableStateOf("") }
-  var descriptionString by remember { mutableStateOf("") }
-  var date by remember { mutableStateOf(LocalDate.now()) }
-  var assignee by remember { mutableStateOf("") }
-  var mutableStatus by remember { mutableStateOf(Status.Pending) }
-  Dialog(onDismissRequest = { balanceDetailedViewModel.cancelEditing() }) {
-    Card(
-      modifier = Modifier
-        .padding(16.dp)
-        .testTag("editDialogBox"),
-      shape = RoundedCornerShape(16.dp),
-    ) {
-      LazyColumn {
-        item {
-          Text("Edit Balance Item", fontSize = 20.sp, modifier = Modifier.padding(16.dp))
-        }
-        // The name box
-        item {
-          OutlinedTextField(
-            modifier = Modifier.padding(8.dp),
-            value = nameString,
-            onValueChange = { nameString = it },
-            label = { Text("Name") },
-            supportingText = {})
-        }
-        // The subcategory selector
-        item {
-          var subcategoryExpanded by remember { mutableStateOf(false) }
-          ExposedDropdownMenuBox(
-            expanded = subcategoryExpanded,
-            onExpandedChange = { subcategoryExpanded = !subcategoryExpanded },
-            modifier = Modifier
-              .testTag("categoryDropdown")
-              .padding(8.dp)
-          ) {
-            OutlinedTextField(
-              value = subCategoryUid,
-              onValueChange = {},
-              label = { Text("SubCategory") },
-              trailingIcon = {
-                ExposedDropdownMenuDefaults.TrailingIcon(expanded = subcategoryExpanded)
-              },
-              readOnly = true,
-              colors = ExposedDropdownMenuDefaults.textFieldColors(),
-              modifier = Modifier
-                .menuAnchor()
-                .clickable { subcategoryExpanded = !subcategoryExpanded })
-            ExposedDropdownMenu(
-              expanded = subcategoryExpanded, onDismissRequest = { subcategoryExpanded = false }) {
-              balanceModel.subCategoryList.forEach { subCat ->
-                DropdownMenuItem(
-                  text = { Text(subCat.name) },
-                  onClick = {
-                    subCategoryUid = subCat.uid
-                    subCategoryName = subCat.name
-                    subcategoryExpanded = false
-                  })
-              }
-            }
-          }
-        }
-        // The receipt selector
-        item {
-          var receiptExpanded by remember { mutableStateOf(false) }
-          ExposedDropdownMenuBox(
-            expanded = receiptExpanded,
-            onExpandedChange = { receiptExpanded = !receiptExpanded },
-            modifier = Modifier
-              .testTag("categoryDropdown")
-              .padding(8.dp)
-          ) {
-            OutlinedTextField(
-              value = receiptUid,
-              onValueChange = {},
-              label = { Text("Receipt") },
-              trailingIcon = {
-                ExposedDropdownMenuDefaults.TrailingIcon(expanded = receiptExpanded)
-              },
-              readOnly = true,
-              colors = ExposedDropdownMenuDefaults.textFieldColors(),
-              modifier = Modifier
-                .menuAnchor()
-                .clickable { receiptExpanded = !receiptExpanded })
-            ExposedDropdownMenu(
-              expanded = receiptExpanded, onDismissRequest = { receiptExpanded = false }) {
-              balanceModel.receiptList.forEach { receipt ->
-                DropdownMenuItem(
-                  text = { Text(receipt.title) },
-                  onClick = {
-                    receiptUid = receipt.uid
-                    receiptName = receipt.title
-                    receiptExpanded = false
-                  })
-              }
-            }
-          }
-        }
-
-        // The amount field
-        item {
-          OutlinedTextField(
-            modifier = Modifier.padding(8.dp),
-            value = amount.toString(),
-            onValueChange = { amount = it.toInt() },
-            label = { Text("Amount") },
-            keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Decimal),
-            supportingText = {})
-        }
-
-        //The TVA box
-        item {
-          var tvaExpanded by remember { mutableStateOf(false) }
-          ExposedDropdownMenuBox(
-            expanded = tvaExpanded,
-            onExpandedChange = { tvaExpanded = !tvaExpanded },
-            modifier = Modifier
-              .testTag("categoryDropdown")
-              .padding(8.dp)
-          ) {
-            OutlinedTextField(
-              value = tvaString,
-              onValueChange = {},
-              label = { Text("Tva") },
-              trailingIcon = {
-                ExposedDropdownMenuDefaults.TrailingIcon(expanded = tvaExpanded)
-              },
-              readOnly = true,
-              colors = ExposedDropdownMenuDefaults.textFieldColors(),
-              modifier = Modifier
-                .menuAnchor()
-                .clickable { tvaExpanded = !tvaExpanded })
-            ExposedDropdownMenu(
-              expanded = tvaExpanded, onDismissRequest = { tvaExpanded = false }) {
-              TVA.entries.forEach { tva ->
-                DropdownMenuItem(
-                  text = { Text(tva.toString()) },
-                  onClick = {
-                    tvaString = tva.rate.toString()
-                    tvaExpanded = false
-                  })
-              }
-            }
-          }
-        }
-
-        //The description field
-        item {
-          OutlinedTextField(
-            modifier = Modifier.padding(8.dp),
-            value = descriptionString,
-            onValueChange = { descriptionString = it },
-            label = { Text("Description") },
-            supportingText = {})
-        }
-        //The date screen
-        item {
-          DatePickerWithDialog(value = "Date", onDateSelect = { newDate ->
-            date = newDate
-          })
-        }
-
-
-        //The assignee field
-        item {
-          OutlinedTextField(
-            modifier = Modifier.padding(8.dp),
-            value = assignee,
-            onValueChange = { assignee = it },
-            label = { Text("Assignee") },
-            supportingText = {})
-        }
-        //The status picker
-        item {
-          var statusExpanded by remember { mutableStateOf(false) }
-          ExposedDropdownMenuBox(
-            expanded = statusExpanded,
-            onExpandedChange = { statusExpanded = !statusExpanded },
-            modifier = Modifier
-              .testTag("categoryDropdown")
-              .padding(8.dp)
-          ) {
-            OutlinedTextField(
-              value = mutableStatus.name,
-              onValueChange = {},
-              label = { Text("Tag") },
-              trailingIcon = {
-                ExposedDropdownMenuDefaults.TrailingIcon(expanded = statusExpanded)
-              },
-              readOnly = true,
-              colors = ExposedDropdownMenuDefaults.textFieldColors(),
-              modifier = Modifier
-                .menuAnchor()
-                .clickable { statusExpanded = !statusExpanded })
-            ExposedDropdownMenu(
-              expanded = statusExpanded, onDismissRequest = { statusExpanded = false }) {
-              Status.entries.forEach { status ->
-                DropdownMenuItem(
-                  text = { Text(status.name) },
-                  onClick = {
-                    mutableStatus = status
-                    statusExpanded = false
-                  })
-              }
-            }
-          }
-        }
-
-
-        //The buttons
-        item {
-          Row(
-            modifier = Modifier
-              .fillMaxWidth()
-              .padding(15.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-          ) {
-            Button(
-              onClick = { balanceDetailedViewModel.cancelEditing() },
-              modifier = Modifier
-                .padding(15.dp)
-                .testTag("editDismissButton"),
-            ) {
-              Text("Dismiss")
-            }
-            Button(
-              onClick = {
-                balanceDetailedViewModel.saveEditing(
-                  BalanceItem(
-                    balance.uid,
-                    nameString,
-                    "",
-                    "",
-                    amount,
-                    TVA.floatToTVA(tvaString.toFloat()),
-                    descriptionString,
-                    date,
-                    assignee,
-                    mutableStatus
-                  )
-                )
-              },
-              modifier = Modifier
-                .padding(15.dp)
-                .testTag("editConfirmButton"),
-            ) {
-              Text("Confirm")
-            }
-          }
-        }
-      }
-    }
-  }
 }
